test: sched: Modify the numa_id from unsigned to int
The numa_id' data type in the struct uacce_dev is int. This
patch modify the data type as same as the uacce_dev.

Signed-ff by: Zhenkun Mi <mi_zhenkun@163.com>

+++ include/sched_sample.h
@@ -28,7 +28,7 @@ typedef int (*user_poll_func)(__u32 pos, __u32 expect, __u32 *count);
  * The shedule indexed mode is NUMA -> MODE -> TYPE -> [BEGIN : END],
  * then select one index from begin to end.
  */
-int sample_sched_fill_data(const struct wd_sched *sched, __u8 numa_id,
+int sample_sched_fill_data(const struct wd_sched *sched, int numa_id,
 			   __u8 mode, __u8 type, __u32 begin, __u32 end);
 
 /**
